refactor: mywishlist and addtowishlist were combined into one endpoint, same for mycart and addtocart. update/delete cart were combined into one end point as well
