+++ src/e_commerce/views/__init__.py
@@ -1,7 +1,4 @@
-from .add_wishlist import AddWishlistView
 from .delete_wishlist import DeleteWishlistView
-from .add_to_cart import AddToCartView
-from .delete_from_cart import DeleteFromCartView
 from .my_wishlist import MyWishListView
 from .my_cart import MyCartView
-from .update_cart import UpdateCartAPIView
+from .update_cart import DetailCartAPIView
